Rename resolver.rs to unnamed.rs in uv-requirements
