+++ crates/uv-requirements/src/lib.rs
@@ -1,14 +1,14 @@
 pub use crate::lookahead::*;
-pub use crate::resolver::*;
 pub use crate::source_tree::*;
 pub use crate::sources::*;
 pub use crate::specification::*;
+pub use crate::unnamed::*;
 
 mod confirm;
 mod lookahead;
 mod pyproject;
-mod resolver;
 mod source_tree;
 mod sources;
 mod specification;
+mod unnamed;
 pub mod upgrade;
